pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeexplain-ai"
version = "1.0.0"
description = "A production-ready static code analysis and explanation engine"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "CodeExplain AI Team" }]
requires-python = ">=3.9"
keywords = ["static-analysis", "ast", "code-explanation", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "rich>=13.7.0",
    "typer>=0.9.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "networkx>=3.3",
    "colorama>=0.4.6",
    "pyfiglet>=1.0.2",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
codeexplain = "cli.main_cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "cli*", "api*", "visualization*", "plugins*", "utils*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ruff]
line-length = 100
target-version = "py39"
